+++ src/common/enums/notification-method.enum.ts
@@ -1,0 +1,6 @@
+export enum NotificationMethod {
+  Sms = 'sms',
+  Email = 'email',
+  Socket = 'socket',
+  fcm = 'fcm',
+}