+++ src/common/interfaces/notifications/index.ts
@@ -2,3 +2,4 @@ export * from './notifier.interface';
 export * from './provider.interface';
 export * from './template.interface';
 export * from './recipients.interface';
+export * from './rules.interface';