+++ src/common/models/notifications/recipient.model.ts
@@ -2,7 +2,7 @@ import { BaseModel } from '../base.model';
 
 export class RecipientModel extends BaseModel<RecipientModel> {
   phone_numbers?: string[];
-  roles?: string[];
+  subjects?: string[];
   user_ids?: string[];
   active?: boolean;
 }