+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Filter pagination
+
 ## 2.4.0 - 2023-07-19
 ### Added
 - one and bulk route